Test Query generation expanded to all date fields and user fields, general clean up of older code and correction on front-end requests/assignments
